requirements.txt
opencv-python==4.7.0.72
pyzbar==0.1.9
Pillow==9.5.0
pywin32==306